requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-heroku==0.3.1
django-jazzmin==3.0.1
django-on-heroku==1.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.2.1
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
marshmallow==3.23.1
moviepy==2.1.1
numpy==2.1.3
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.2
stripe==11.3.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
whitenoise==6.8.2
